+++ CHANGES.md
@@ -41,6 +41,7 @@ Changes in 10.0
 - Shorten Referral email subject in EmailReferral.fnc.php
 - Use plural wise ngettext() for "No %s were found." in FinalGrades.php, GradeBreakdown.php, ReportCardComments.php, ReportCards.php & Transcripts.php
 - Force result text to lowercase for "No %s were found." in ListOutput.fnc.php, FinalGrades.php, GradeBreakdown.php, ReportCardComments.php, ReportCards.php & Transcripts.php
+- Prevent admin from removing own access to User Profiles program in Profiles.php
 
 Changes in 9.3
 --------------